Starting refactoring work
