+++ auth.py
@@ -2,10 +2,11 @@ import hashlib
 import os
 from functools import wraps
 
-from flask import request, jsonify
+from flask import request
 from error_handlers import _make_response
 
 FORTNITE_REPLAY_ELIM_API_AUTH_DIGEST = os.getenv("FORTNITE_REPLAY_ELIM_API_AUTH_DIGEST")
+
 
 def validate(f):
     """ Validate headers for incoming requests
@@ -15,7 +16,7 @@ def validate(f):
         # Validate header
         if "API-TOKEN" not in request.headers:
             return _make_response("Authentication header not provided", 400)
-        
+
         req_api_token = request.headers.get("API-TOKEN", "")
         req_api_digest = hashlib.sha512(req_api_token.encode()).hexdigest()
         if req_api_digest != FORTNITE_REPLAY_ELIM_API_AUTH_DIGEST:
@@ -23,4 +24,3 @@ def validate(f):
 
         return f(*args, **kwargs)
     return validate
-
